Consider only the out-in MEP to wait for synchronous response
This reverts commits d50ec6baf1bf8a629fb564c6d408fe1f3a354594
and 8aae45971be2f98ed9d1c9a66c7cc34c5a7f68ec which were intended to
fix the issue [1] with pr [2]. The issue [1] is not reproducible
anymore according to the logic in the JMSSender class in the master
branch. In the latest EI pack it can be observed that the "ReplyTo"
header is getting stamped in tenant environment even without the fix
in pr [2]. Therefore reverting the fix should be fine.

Reverting that fix resolves the issue [3] which is occurred due to
waiting for a response in the response path for the message which has
the MEP (Message exchange pattern) of in-out.

Fixes wso2/product-ei#2333

[1] - https://github.com/wso2/product-ei/issues/1036
[2] - https://github.com/wso2/wso2-axis2/pull/127
[3] - https://github.com/wso2/product-ei/issues/2333